+++ .github/workflows/build.yml
@@ -1,39 +1,40 @@
-# Automatically build the project and run any configured tests for every push
-# and submitted pull request. This can help catch issues that only occur on
-# certain platforms or Java versions, and provides a first line of defence
-# against bad commits.
+name: Build
 
-name: build
-on: [pull_request, push]
+on: [ pull_request, push, workflow_dispatch ]
 
 jobs:
-  build:
-    strategy:
-      matrix:
-        # Use these Java versions
-        java: [
-          17,    # Current Java LTS & minimum supported by Minecraft
-        ]
-        # and run on both Linux and Windows
-        os: [ubuntu-20.04, windows-2022]
-    runs-on: ${{ matrix.os }}
+  Build:
+    runs-on: ubuntu-22.04
+
     steps:
-      - name: checkout repository
-        uses: actions/checkout@v2
-      - name: validate gradle wrapper
-        uses: gradle/wrapper-validation-action@v1
-      - name: setup jdk ${{ matrix.java }}
-        uses: actions/setup-java@v1
+      - uses: actions/checkout@v3
+      - name: Get github short hash
+        id: github_short_hash
+        run: echo "result=$(git rev-parse --short "${{ github.sha }}")" >> $GITHUB_OUTPUT
+      - name: Set up JDK 17
+        uses: actions/setup-java@v3
         with:
-          java-version: ${{ matrix.java }}
-      - name: make gradle wrapper executable
-        if: ${{ runner.os != 'Windows' }}
-        run: chmod +x ./gradlew
-      - name: build
-        run: ./gradlew build
-      - name: capture build artifacts
-        if: ${{ runner.os == 'Linux' && matrix.java == '17' }} # Only upload artifacts built from latest java on one OS
-        uses: actions/upload-artifact@v2
+          distribution: 'temurin'
+          java-version: 17
+      - name: Cache
+        uses: actions/cache@v2
         with:
-          name: Artifacts
-          path: build/libs/
+          path: |
+            ~/.gradle/caches
+            ~/.gradle/wrapper
+            .gradle
+          key: ${{ runner.os }}-gradle-${{ hashFiles('**/*.gradle*', '**/gradle-wrapper.properties', 'build.gradle') }}
+      - name: Build with Gradle
+        run: |
+          chmod +x gradlew
+          ./gradlew build
+      - name: Upload artifacts (fabric)
+        uses: actions/upload-artifact@v3
+        with:
+          name: 1.18.x-${{ steps.github_short_hash.outputs.result }}-fabric
+          path: ${{ github.workspace }}/fabric/build/libs
+      - name: Upload artifacts (forge)
+        uses: actions/upload-artifact@v3
+        with:
+          name: 1.18.x-${{ steps.github_short_hash.outputs.result }}-forge
+          path: ${{ github.workspace }}/forge/build/libs
