+++ .gitignore
@@ -1,33 +1,19 @@
-# gradle
-
-.gradle/
 build/
-out/
-classes/
-
-# eclipse
-
-*.launch
-
-# idea
-
-.idea/
-*.iml
 *.ipr
+run/
 *.iws
-
-# vscode
-
-.settings/
-.vscode/
+out/
+*.iml
+.gradle/
+output/
 bin/
+libs/
+
 .classpath
 .project
-
-# macos
-
-*.DS_Store
-
-# fabric
-
-run/
+.idea/
+classes/
+.metadata
+.vscode
+.settings
+*.launch
