+++ common/src/main/resources/assets/minegpt/lang/en_us.json
