+++ common/src/main/resources/assets/minegpt/lang/zh_cn.json
