+++ fabric-old/gradle.properties
@@ -1,0 +1,16 @@
+# Done to increase the memory available to gradle.
+org.gradle.jvmargs=-Xmx1G
+
+# Fabric Properties
+	# check these on https://fabricmc.net/develop
+	minecraft_version=1.18.2
+    yarn_mappings=1.18.2+build.4
+    loader_version=0.14.17
+
+# Mod Properties
+	mod_version = 1.1.3
+	maven_group = team.teampotato
+	archives_base_name = fabric-MineGPT
+
+# Dependencies
+    fabric_version=0.75.1+1.18.2
