+++ fabric-old/src/main/java/team/teampotato/chatgpt/Command.java
