+++ fabric-old/src/main/resources/assets/minegpt/lang/en_us.json
@@ -1,0 +1,7 @@
+{
+  "minegpt.set.configapikey": "OpenAI API key saved.",
+  "minegpt.set.default": "The default endpoint has been updated in %s",
+  "minegpt.error.saveconfig": "Failed to save config file.",
+  "minegpt.error.apikey": "Invalid API key!",
+  "minegpt.reload.config": "The configuration file has been reloaded."
+}
