+++ fabric-old/src/main/resources/fabric.mod.json
