+++ fabric/src/main/java/team/teampotato/chatgpt/fabric/MineGPTFabric.java
@@ -1,0 +1,11 @@
+package team.teampotato.chatgpt.fabric;
+
+import team.teampotato.chatgpt.MineGPT;
+import net.fabricmc.api.ModInitializer;
+
+public class MineGPTFabric implements ModInitializer {
+    @Override
+    public void onInitialize() {
+        MineGPT.init();
+    }
+}
