+++ forge/src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+  "pack": {
+    "description": "MineGPT",
+    "pack_format": 8
+  }
+}
