+++ gradle.properties
@@ -1,16 +1,20 @@
-# Done to increase the memory available to gradle.
-org.gradle.jvmargs=-Xmx1G
+org.gradle.jvmargs=-Xmx2048M
 
-# Fabric Properties
-	# check these on https://fabricmc.net/develop
-	minecraft_version=1.18.2
-    yarn_mappings=1.18.2+build.4
-    loader_version=0.14.17
+minecraft_version=1.18.2
+enabled_platforms=fabric,forge
 
-# Mod Properties
-	mod_version = 1.1.3
-	maven_group = team.teampotato
-	archives_base_name = fabric-MineGPT
+yarn_mappings=1.18.2+build.4
 
-# Dependencies
-    fabric_version=0.75.1+1.18.2
+archives_base_name=MineGPT
+mod_version=1.1.4
+maven_group=team.teampotato
+
+architectury_version=4.11.91
+
+fabric_loader_version=0.14.19
+fabric_api_version=0.76.0+1.18.2
+
+forge_version=1.18.2-40.2.1
+
+
+
